bench/tpcc: consolidate utilities in the same file
Release note: None
